cap-03/SRP: Refatora a nivel de metodo. Aplica o SRP a nivel de metodo nas classes RenderizadorMDParaHTML e LeitorOpcoesCLI. Cria metodos privados para separar a logica e o passo a passo da logica das classes.
