+++ src/main/java/cotuba/cli/Main.java
@@ -1,4 +1,6 @@
-package cotuba;
+package cotuba.cli;
+
+import cotuba.application.Cotuba;
 
 import java.nio.file.Path;
