+++ src/main/java/cotuba/epub/GeradorEPUB.java
@@ -1,5 +1,7 @@
-package cotuba;
+package cotuba.epub;
 
+import cotuba.domain.Capitulo;
+import cotuba.domain.Ebook;
 import nl.siegmann.epublib.domain.Book;
 import nl.siegmann.epublib.domain.Resource;
 import nl.siegmann.epublib.epub.EpubWriter;
