+++ src/main/java/cotuba/pdf/GeradorPDF.java
@@ -1,4 +1,4 @@
-package cotuba;
+package cotuba.pdf;
 
 import com.itextpdf.html2pdf.HtmlConverter;
 import com.itextpdf.kernel.pdf.PdfDocument;
@@ -8,6 +8,8 @@ import com.itextpdf.layout.element.AreaBreak;
 import com.itextpdf.layout.element.IBlockElement;
 import com.itextpdf.layout.element.IElement;
 import com.itextpdf.layout.property.AreaBreakType;
+import cotuba.domain.Capitulo;
+import cotuba.domain.Ebook;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
