+++ src/scripts/cotuba.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-java -cp "libs/*" cotuba.Main "$@"
+java -cp "libs/*" cotuba.cli.Main "$@"
